requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.4
flake8==7.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
openpyxl==3.1.5
greenlet==3.3.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
nulltype==2.3.1
plaid-python==24.0.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
WTForms==3.1.2
gunicorn==22.0.0

# Windows-specific dependency
pywin32==306; platform_system == "Windows"
